+++ jwst/associations/lib/rules_level3_base.py
@@ -21,7 +21,6 @@ from jwst.associations.lib.utilities import (
     is_iterable
 )
 from jwst.associations.exceptions import (
-    #AssociationNotAConstraint,
     AssociationNotValidError,
 )
 from jwst.associations.lib.acid import ACID
@@ -33,13 +32,14 @@ from jwst.associations.lib.counter import Counter
 from jwst.associations.lib.dms_base import (
     _EMPTY,
     ACQ_EXP_TYPES,
+    Constraint_TargetAcq,
     CORON_EXP_TYPES,
     DMSAttrConstraint,
     DMSBaseMixin,
     IMAGE2_SCIENCE_EXP_TYPES,
     IMAGE2_NONSCIENCE_EXP_TYPES,
     SPEC2_SCIENCE_EXP_TYPES,
-    )
+)
 from jwst.associations.lib.format_template import FormatTemplate
 from jwst.associations.lib.member import Member
 
@@ -817,7 +817,6 @@ class Constraint_Target(DMSAttrConstraint):
             )
 
 
-
 # -----------
 # Base Mixins
 # -----------
@@ -829,12 +828,7 @@ class AsnMixin_Science(DMS_Level3_Base):
         # Setup target acquisition inclusion
         constraint_acqs = Constraint(
             [
-                DMSAttrConstraint(
-                    name='acq_exp',
-                    sources=['exp_type'],
-                    value='|'.join(ACQ_EXP_TYPES),
-                    force_unique=False
-                ),
+                Constraint_TargetAcq(),
                 DMSAttrConstraint(
                     name='acq_obsnum',
                     sources=['obs_num'],
@@ -884,12 +878,7 @@ class AsnMixin_BkgScience(DMS_Level3_Base):
         # Setup target acquisition inclusion
         constraint_acqs = Constraint(
             [
-                DMSAttrConstraint(
-                    name='acq_exp',
-                    sources=['exp_type'],
-                    value='|'.join(ACQ_EXP_TYPES),
-                    force_unique=False
-                ),
+                Constraint_TargetAcq(),
                 DMSAttrConstraint(
                     name='acq_obsnum',
                     sources=['obs_num'],
@@ -960,14 +949,12 @@ class AsnMixin_AuxData(AsnMixin_Science):
         exposure_type : 'target_acquisition'
             Returns target_acquisition for mir_tacq
         """
-        try:
-            exp_type = item['exp_type']
-            if any(x == exp_type for x in ACQ_EXP_TYPES):
-                return 'target_acquisition'
-        except KeyError:
-            raise LookupError('Exposure type cannot be determined')
-
+        NEVER_CHANGE = ['target_acquisition']
+        exp_type = super().get_exposure_type(item, default=default)
+        if exp_type in NEVER_CHANGE:
+            return exp_type
         return 'science'
+
     def _init_hook(self, item):
         """Post-check and pre-add initialization"""